Change SuspendTime configuration parameter from 16-bits to 32-bits.

+++ src/common/slurm_protocol_api.h
@@ -111,7 +111,7 @@ uint16_t slurm_get_resume_timeout(void);
 /* slurm_get_suspend_time
  * RET SuspendTime value from slurm.conf
  */
-uint16_t slurm_get_suspend_time(void);
+uint32_t slurm_get_suspend_time(void);
 
 /* slurm_get_complete_wait
  * RET CompleteWait value from slurm.conf
